more progress in direct orderbook websocket subscriber

+++ src/main/java/com/cb/model/kraken/ws/KrakenStatusUpdateData.java
@@ -4,13 +4,15 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @ToString
 public class KrakenStatusUpdateData {
 
     private String api_version;
-    private long connection_id;
+    private BigInteger connection_id;
     private String system;
     private String version;
